% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridref_to_xy.R
\name{gridref_to_xy}
\alias{gridref_to_xy}
\title{Convert Grid Ref to XY}
\usage{
gridref_to_xy(x)
}
\arguments{
\item{x}{A vector of ordinance survey grid references such as 'TA 123 231'}
}
\value{
A data.frame of eastings and northings coordinates and their resolution
}
\description{
This function converts ordinance survey grid references to eastings northing
}
