% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osgb_lookup.R
\name{osgb_lookup}
\alias{osgb_lookup}
\title{OS GB Look up table}
\usage{
osgb_lookup()
}
\value{
A \code{data.frame} with 3 columns, the letters identifying each square and their x and y min coordinates
}
\description{
This function creates a lookup table of os grid square letters and numbers
}
\keyword{grid}
\keyword{look}
\keyword{osgb}
\keyword{reference}
\keyword{up}
