% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square_lookup.R
\docType{data}
\name{square_lookup}
\alias{square_lookup}
\title{Square coordinates of the British National Grid}
\format{
A data frame with 104 rows and 3 variables:
\describe{
\item{x}{xmin of grid square}
\item{y}{ymin of grid square}
\item{square_letters}{the two letter code for each grid square}
}
}
\usage{
square_lookup
}
\description{
A dataset containing the letter codes, xmin and ymin coordinates of squares on the british naitonal grid
}
\keyword{datasets}
