% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xy_to_gridref.R
\name{xy_to_gridref}
\alias{xy_to_gridref}
\title{XY to Grid Ref}
\usage{
xy_to_gridref(x, y, digits = 10)
}
\arguments{
\item{x}{a vector of x coordinates (must be epsg 27700)}

\item{y}{a vector of y coordinates (must be epsg 27700)}

\item{digits}{the number of digits to return (maximum value is 10)}
}
\value{
a vector of osgb grid references
}
\description{
Convert x and y coordinates to an osgb grid reference
}
